Add implementation of IPipelineCreationAPIDispatcher
Because slang-gfx doesn't provide a fine granularity APIs such
that we can time the spirv generation phase and Vulkan pipeline
creation phase, we have to use gfx::IPipelineCreationAPIDispatcher
callback interface so that the gfx will call it when creating pipeline.

We simply call Vulkan pipeline creation inside the callback function and
measure the time so that we can get the time of Vulkan driver compiling
the spirv code.